requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
idna==3.7
oauthlib==3.2.2
openpyxl==3.1.2
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
urllib3==2.2.1
